Move the session tracking logic into its own package
